Update markdown configuration and styling
This commit updates the markdown configuration in the `config.toml` file to enable code highlighting and sets the highlight theme to "charcoal". Additionally, it adds a CSS rule in the `_custom.scss` file to add a margin-bottom of 2.5rem to the `div` element with the attribute `itemprop="articleBody"`. Lastly, it modifies the `macros.html` template to display the publication date in the format "YYYY-MM-DD HH:MM UTC".